(BSR)[PRO] refactor: move components out of index
