+++ pro/src/ui-kit/form/SelectAutoComplete2/Toggle/index.tsx
@@ -1,37 +1,1 @@
-import cx from 'classnames'
-import React from 'react'
-
-import Icon from 'ui-kit/Icon/Icon'
-
-import styles from './Toggle.module.scss'
-
-export interface ToggleProps {
-  disabled: boolean
-  isOpen: boolean
-  toggleField: () => void
-}
-
-const Toggle = ({
-  disabled,
-  isOpen,
-  toggleField,
-}: ToggleProps): JSX.Element => {
-  return (
-    <button
-      onClick={toggleField}
-      className={cx(styles['dropdown-indicator'], {
-        [styles['dropdown-indicator-is-closed']]: !isOpen,
-      })}
-      type="button"
-      disabled={disabled}
-    >
-      <Icon
-        svg="open-dropdown"
-        alt={`${isOpen ? 'Masquer' : 'Afficher'} les options`}
-        className={styles['dropdown-indicator-icon']}
-      />
-    </button>
-  )
-}
-
-export default Toggle
+export { default } from './Toggle'
